[SPARK-44993][CORE] Move compareChecksums from ShuffleChecksumTestHelpe to ShuffleChecksumUtils
